requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.2.0
SQLAlchemy==2.0.35
python-dotenv==1.0.1
WTForms==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.4
gunicorn==22.0.0